+++ examples/with-next-custom-button/src/wagmi.ts
@@ -11,6 +11,8 @@ import {
 
 export const config = getDefaultConfig({
   appName: 'RainbowKit demo',
+  appUrl: 'https://rainbow.me',
+  appIcon: 'https://framerusercontent.com/images/Hml6PtJwt03gwFtTRYmbpo7EarY.png',
   projectId: 'YOUR_PROJECT_ID',
   chains: [
     mainnet,