+++ examples/with-next-rainbow-button/src/wagmi.ts
@@ -6,6 +6,8 @@ export const config = createConfig({
   connectors: [
     rainbowConnector({
       appName: 'RainbowKit demo',
+      appUrl: 'https://rainbow.me',
+      appIcon: 'https://framerusercontent.com/images/Hml6PtJwt03gwFtTRYmbpo7EarY.png',
       projectId: 'YOUR_PROJECT_ID',
     }),
   ],